requirements.txt
discord.py
python-dotenv
groq
google-generativeai
